+++ src/components/flat-table/flat-table.component.js
@@ -4,11 +4,13 @@ import { StyledFlatTableWrapper, StyledFlatTable } from './flat-table.style';
 
 const FlatTable = ({
   children,
-  hasStickyHead
+  hasStickyHead,
+  colorTheme
 }) => {
   return (
     <StyledFlatTableWrapper
       hasStickyHead={ hasStickyHead }
+      colorTheme={ colorTheme }
     >
       <StyledFlatTable data-component='flat-table'>
         { children }
@@ -21,7 +23,13 @@ FlatTable.propTypes = {
   /** FlatTableHead and FlatTableBody */
   children: PropTypes.node.isRequired,
   /** If true, the header does not scroll with the content */
-  hasStickyHead: PropTypes.bool
+  hasStickyHead: PropTypes.bool,
+  /** `FlatTable` color theme */
+  colorTheme: PropTypes.oneOf(['light', 'transparent-base', 'transparent-white', 'dark'])
+};
+
+FlatTable.defaultProps = {
+  colorTheme: 'dark'
 };
 
 export default FlatTable;